auto: 17:01 auto save

+++ yanguo基于校赛改了的，能跑副本/project/code/Camera/Draw_Line.h
@@ -5,9 +5,9 @@
 
 void Patch_line_Left(int16 zuobiao_botton[2], int16 zuobiao_top[2]);
 void Patch_line_Right(int16 zuobiao_botton[2], int16 zuobiao_top[2]);
-void Add_mid_left(int start, int end);
-void Patch_line_Right_Init(int start, int end);
-void Patch_line_Left_Init(int start, int end);
+void Add_mid_left(int16 start, int16 end);
+void Patch_line_Right_Init(int16 start, int16 end);
+void Patch_line_Left_Init(int16 start, int16 end);
 void Right_Patch_Init(void);
 void Left_Patch_Init(void);
